WIP: gb: Standard-compliant Static Gb over IP
The current implementation of the Gb over IP in Osmocom is derived
from the ip.access implementation of it, which is a "weird combination
of Gb over FR and Gb over IP" in that it runs over UDP/IP but uses
procedures like NS-RESET and NS-BLOCK which are only specified
for Gb over FR.

This makes it impossible to use OsmoPCU and OsmoGbProxy with
standard SGSNs like from Huawei in "Static IP-GB" mode
(i.e. without SNS procedure).

This patch is a hack to remove NS-RESET procedure and use NS-ALIVE
to setup an NS link and keep it open.

One could argue that SNS is a much better way and should be used
instead of the old "static" config but in pratical installations
you do ant to use osmo-gb-proxy to aggregate multiple links and
it doesn't support SNS yet. So using this hack is the easiest way
to get multiple OsmoPCU's to connect to a standrd-compliant SGSN.

NOTE: This patch is a quick hack and lacks any unit-testing or
configuration.

Change-Id: I1b1b28913488a40e4fceb65e646c3d89e8a431a4